refactor: make RequiredDAAs separate component
