+++ src/pages/dar_application/DataAccessAgreements.jsx
@@ -2,10 +2,11 @@ import React, { useEffect, useState } from 'react';
 import {isNil} from 'lodash/fp';
 import { Notifications } from '../../libs/utils';
 import { DAA } from '../../libs/ajax/DAA';
+import RequiredDAAs from './RequiredDAAs';
 
 import './dar_application.css';
 
-export default function DataUseAgreements(props) {
+export default function DataAccessAgreements(props) {
 
   const {
     save,
@@ -30,42 +31,6 @@ export default function DataUseAgreements(props) {
     };
     init();
   }, []);
-
-  const RequiredDAAs = () => {
-    const fileNames = new Set();
-    const daaDivs = datasets.map((dataset) => {
-      const datasetDacId = dataset.dacId;
-      if (!datasetDacId) {
-        return <div key={dataset.id}></div>;
-      }
-      const daa = daas.find((daa) => daa.dacs.some((d) => d.dacId === datasetDacId));
-      const id = daa.daaId;
-      const fileName = daa.file.fileName.split('.')[0];
-      if (fileNames.has(fileName)) {
-        return <div key={id-dataset.name}></div>;
-      }
-      fileNames.add(fileName);
-      return (
-        <div key={id}>
-          {DAADownload(id, fileName)}
-        </div>
-      );
-    });
-    if (fileNames.size === 0) {
-      return (
-        <div></div>
-      );
-    } else {
-      return (
-        <div>
-          <h3>By submitting this data access request and in accordance with your Institution’s issuance of Library Cards to you for the agreement(s) below.</h3>
-          <div className="flex flex-row" style={{ justifyContent: 'flex-start' }}>
-            {daaDivs}
-          </div>
-        </div>
-      );
-    }
-  };
 
   const DAADownload = (id, fileName) => {
     return (
@@ -103,7 +68,7 @@ export default function DataUseAgreements(props) {
         </ol>
       </div>
 
-      <RequiredDAAs/>
+      <RequiredDAAs datasets={datasets} daas={daas} daaDownload={DAADownload}/>
 
       <div className="flex flex-row" style={{ justifyContent: 'around', paddingTop: '4rem' }}>
         <div className="flex flex-row" style={{ justifyContent: 'flex-start' }}>
